增加宏

+++ components/net/at_frame/at_main.c
@@ -283,7 +283,7 @@ int32_t at_cmd_multi_suffix(const int8_t *cmd, int  len, at_cmd_info_s *cmd_info
     memset(&listener, 0, sizeof(listener));
     listener.cmd_info = *cmd_info;
     print_len = ((cmd_info->resp_buf && cmd_info->resp_len) ? (int)*(cmd_info->resp_len) : -1);
-    AT_LOG("cmd:%s len %d, %p,%d", cmd, print_len,cmd_info->resp_buf,cmd_info->resp_len);
+    AT_LOG("cmd:%s len %d, %p,%d", cmd, print_len,cmd_info->resp_buf,(int)cmd_info->resp_len);
 
     LOS_MuxPend(at.trx_mux, LOS_WAIT_FOREVER);
